chore: renamed Calcver.Git.Cli to Calcver.Cli
resolves #25
