+++ appveyor.yml
@@ -10,8 +10,8 @@ branches:
   - /v\d+\.\d+/
 before_build:
 - ps: $env:versionSuffix = if ($env:APPVEYOR_PULL_REQUEST_NUMBER) { "pr$env:APPVEYOR_PULL_REQUEST_NUMBER.$env:APPVEYOR_BUILD_ID" } else { "$env:APPVEYOR_BUILD_ID" }
-- ps: dotnet publish "$env:APPVEYOR_BUILD_FOLDER\src\Calcver.Git.Cli" -c release -o "$env:APPVEYOR_BUILD_FOLDER\build"
-- ps: $env:customnugetversion = dotnet "$env:APPVEYOR_BUILD_FOLDER\build\Calcver.Git.Cli.dll" version "$env:APPVEYOR_BUILD_FOLDER" "$env:versionSuffix"
+- ps: dotnet publish "$env:APPVEYOR_BUILD_FOLDER\src\Calcver.Cli" -c release -o "$env:APPVEYOR_BUILD_FOLDER\build"
+- ps: $env:customnugetversion = dotnet "$env:APPVEYOR_BUILD_FOLDER\build\Calcver.Cli.dll" version "$env:APPVEYOR_BUILD_FOLDER" "$env:versionSuffix"
 - ps: Update-AppveyorBuild -Version $env:customnugetversion
 - choco install opencover.portable
 - choco install codecov
