+++ src/Calcver.Cli/Calcver.Cli.csproj
